Move tests for core variable processing
