+++ seaborn/conftest.py
@@ -1,5 +1,7 @@
 import numpy as np
+import pandas as pd
 import matplotlib.pyplot as plt
+
 import pytest
 
 
@@ -12,3 +14,105 @@ def close_figs():
 @pytest.fixture(autouse=True)
 def random_seed():
     np.random.seed(47)
+
+
+@pytest.fixture
+def wide_df(self):
+
+    columns = list("abc")
+    index = pd.Int64Index(np.arange(10, 50, 2), name="wide_index")
+    values = np.random.randn(len(index), len(columns))
+    return pd.DataFrame(values, index=index, columns=columns)
+
+
+@pytest.fixture
+def wide_array(self, wide_df):
+
+    return wide_df.to_numpy()
+
+
+@pytest.fixture
+def flat_array(self, flat_series):
+
+    return flat_series.to_numpy()
+
+
+@pytest.fixture
+def flat_list(self, flat_series):
+
+    return flat_series.to_list()
+
+
+@pytest.fixture
+def wide_list_of_series(self):
+
+    return [pd.Series(np.random.randn(20), np.arange(20), name="a"),
+            pd.Series(np.random.randn(10), np.arange(5, 15), name="b")]
+
+
+@pytest.fixture
+def wide_list_of_arrays(self, wide_list_of_series):
+
+    return [s.to_numpy() for s in wide_list_of_series]
+
+
+@pytest.fixture
+def wide_list_of_lists(self, wide_list_of_series):
+
+    return [s.to_list() for s in wide_list_of_series]
+
+
+@pytest.fixture
+def long_df(self):
+
+    n = 100
+    rs = np.random.RandomState()
+    df = pd.DataFrame(dict(
+        x=rs.randint(0, 20, n),
+        y=rs.randn(n),
+        a=np.take(list("abc"), rs.randint(0, 3, n)),
+        b=np.take(list("mnop"), rs.randint(0, 4, n)),
+        c=np.take(list([0, 1]), rs.randint(0, 2, n)),
+        d=np.repeat(np.datetime64('2005-02-25'), n),
+        s=np.take([2, 4, 8], rs.randint(0, 3, n)),
+        f=np.take(list([0.2, 0.3]), rs.randint(0, 2, n)),
+    ))
+    df["s_cat"] = df["s"].astype("category")
+    return df
+
+
+@pytest.fixture
+def repeated_df(self):
+
+    n = 100
+    rs = np.random.RandomState()
+    return pd.DataFrame(dict(
+        x=np.tile(np.arange(n // 2), 2),
+        y=rs.randn(n),
+        a=np.take(list("abc"), rs.randint(0, 3, n)),
+        u=np.repeat(np.arange(2), n // 2),
+    ))
+
+
+@pytest.fixture
+def missing_df(self):
+
+    n = 100
+    rs = np.random.RandomState()
+    df = pd.DataFrame(dict(
+        x=rs.randint(0, 20, n),
+        y=rs.randn(n),
+        a=np.take(list("abc"), rs.randint(0, 3, n)),
+        b=np.take(list("mnop"), rs.randint(0, 4, n)),
+        s=np.take([2, 4, 8], rs.randint(0, 3, n)),
+    ))
+    for col in df:
+        idx = rs.permutation(df.index)[:10]
+        df.loc[idx, col] = np.nan
+    return df
+
+
+@pytest.fixture
+def null_column(self):
+
+    return pd.Series(index=np.arange(20), dtype='float64')
